+++ robospice-cache-parent/robospice-cache-test/src/main/java/com/octo/android/robospice/persistence/file/AbstractInFileObjectPersisterTest.java
@@ -10,6 +10,7 @@ import android.test.suitebuilder.annotation.MediumTest;
 @MediumTest
 public abstract class AbstractInFileObjectPersisterTest extends InstrumentationTestCase {
 
+    private static final int TEST_DATE_IN_CACHE = 5000;
     private static final long TEST_EXPIRATION_DURATION = 1000;
     protected InFileObjectPersister<Object> inFileObjectPersister;
 
@@ -113,4 +114,32 @@ public abstract class AbstractInFileObjectPersisterTest extends InstrumentationT
         assertFalse(inFileObjectPersister.isDataInCache(cacheKey, Long.MAX_VALUE));
     }
 
+    public void testGetDateOfDataInCache_when_there_is_some_data_in_cache(Object data, Object cacheKey) throws Exception {
+        // given
+        inFileObjectPersister.saveDataToCacheAndReturnData(data, cacheKey);
+
+        // when
+        inFileObjectPersister.getCacheFile(cacheKey).setLastModified(TEST_DATE_IN_CACHE);
+
+        // then
+        assertEquals(TEST_DATE_IN_CACHE, inFileObjectPersister.getCreationDateInCache(cacheKey));
+
+    }
+
+    public void testGetDateOfDataInCache_when_there_is_no_data_in_cache(Object data, Object cacheKey) throws Exception {
+        // given
+        inFileObjectPersister.removeDataFromCache(cacheKey);
+
+        // when
+
+        // then
+        try {
+            inFileObjectPersister.getCreationDateInCache(cacheKey);
+            fail("Should have thrown an exception");
+        } catch (Exception ex) {
+            // nothing
+            assertTrue(true);
+        }
+
+    }
 }