+++ robospice-core-parent/robospice-core-test/src/main/java/com/octo/android/robospice/stub/StringPersisterStub.java
@@ -51,7 +51,7 @@ public class StringPersisterStub extends ObjectPersister<String> {
 
     @Override
     public long getCreationDateInCache(Object cacheKey) throws CacheLoadingException {
-        return 0;
+        throw new CacheLoadingException("This persisters stores no data.");
     }
 
     @Override